+++ stories/User Interaction/Date and Time Input/YearMonthPicker.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/nextjs'
 import { action } from 'storybook/actions'
 import { YearMonthPicker } from '@/src/components/user-interaction/date/YearMonthPicker'
-import { addDuration, subtractDuration } from '@/src/utils/date'
+import { DateUtils } from '@/src/utils/date'
 
 const meta = {
   component: YearMonthPicker,
@@ -13,8 +13,8 @@ type Story = StoryObj<typeof meta>;
 export const yearMonthPicker: Story = {
   args: {
     initialValue: new Date(),
-    start: subtractDuration(new Date(), { years: 50 }),
-    end: addDuration(new Date(), { years: 50 }),
+    start: DateUtils.subtractDuration(new Date(), { years: 50 }),
+    end: DateUtils.addDuration(new Date(), { years: 50 }),
     onValueChange: action('onValueChange'),
     onEditComplete: action('onEditComplete'),
   },