+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.2] - 2026-02-13
+
+### Added
+- `useUpdatingDateString` to show absolute or relative Date strings
+- time input to `DateTimeInput`
+
+### Changed
+- `DateTimeInput` to allow for writing input
+
 ## [0.8.1] - 2026-02-08
 
 ### Fixed